requirements.txt
torch>=2.0.0
datasets>=2.0.0
tokenizers>=0.13.0
numpy>=1.21.0
matplotlib>=3.5.0
tqdm>=4.64.0
PyYAML>=6.0
pandas>=1.5.0